+++ src/firelord.ts
@@ -1,21 +1,23 @@
 import { FirelordFirestore } from './firelordFirestore'
+import { DeepRequired } from 'ts-essentials'
 
 export type OmitKeys<T, K extends keyof T> = Omit<T, K>
 
 export type RemoveArray<T extends unknown[]> = T extends (infer A)[] ? A : never
 
-export type DeepRequired<T> = Required<{
-	[P in keyof T]: T[P] extends Record<string, unknown>
-		? DeepRequired<T[P]>
-		: T[P]
-}>
+// use ts-essentials DeepRequired because it has cleaner typescript hint
+// export type DeepRequired<T> = Required<{
+// 	[P in keyof T]: T[P] extends Record<string, unknown>
+// 		? DeepRequired<T[P]>
+// 		: T[P] extends (infer A)[]
+// 		? DeepRequired<A>[]
+// 		: T[P]
+// }>
 
 export type ExcludePropertyKeys<A, U = undefined> = string &
 	{
 		[P in keyof A]: A[P] extends U ? never : P
 	}[keyof A]
-
-type IncludeKeys<T, K extends keyof T> = { [Y in K]: T[Y] }
 
 type DistributeNoUndefined<T, K> = T extends undefined
 	? T
@@ -27,12 +29,23 @@ export type PartialNoImplicitUndefinedAndNoExtraMember<
 	L extends { [index: string]: unknown },
 	T extends Partial<L>
 > = keyof T extends keyof L
-	? IncludeKeys<
-			{
-				[K in keyof L]: DistributeNoUndefined<L[K], T[K]>
-			},
-			keyof L & keyof T
-	  >
+	? {
+			[K in keyof L & keyof T]: L[K] extends Record<string, unknown>
+				? T[K] extends Partial<L[K]>
+					? PartialNoImplicitUndefinedAndNoExtraMember<L[K], T[K]>
+					: never
+				: L[K] extends (infer A)[]
+				? T[K] extends (infer B)[]
+					? A extends Record<string, unknown>
+						? B extends Partial<A>
+							? PartialNoImplicitUndefinedAndNoExtraMember<A, B>[]
+							: never
+						: B extends A
+						? DistributeNoUndefined<A, B>[]
+						: never[]
+					: never
+				: DistributeNoUndefined<L[K], T[K]>
+	  }
 	: never
 
 export namespace Firelord {
@@ -50,7 +63,7 @@ export namespace Firelord {
 	// https://javascript.plainenglish.io/using-firestore-with-more-typescript-8058b6a88674
 	type DeepKey<T, K extends keyof T> = K extends string
 		? // Date, timestamp and geo point will never extends Record<string, unknown>, so we dont need these lines
-		  // ! however why these few line cause <Type instantiation is excessively deep and possibly infinite> if enabled? WHY?
+		  // ! however why these lines trigger <Type instantiation is excessively deep and possibly infinite> error if enabled? how is this possible?
 		  // T[K] extends
 		  // 		| Date
 		  // 		| FirelordFirestore.Timestamp
@@ -85,7 +98,9 @@ export namespace Firelord {
 
 	type ArrayWriteConverter<T> = T extends (infer A)[]
 		? ArrayWriteConverter<A>[]
-		: T extends FirelordFirestore.Timestamp | Date | ServerTimestamp
+		: T extends ServerTimestamp
+		? never
+		: T extends FirelordFirestore.Timestamp | Date
 		? FirelordFirestore.Timestamp | Date
 		: T extends Record<string, unknown>
 		? {
@@ -95,7 +110,7 @@ export namespace Firelord {
 
 	type ReadConverter<T> = T extends (infer A)[]
 		? ReadConverter<A>[]
-		: T extends ServerTimestamp | Date
+		: T extends ServerTimestamp | Date | FirelordFirestore.Timestamp
 		? FirelordFirestore.Timestamp
 		: T extends Record<string, unknown>
 		? {
@@ -123,26 +138,6 @@ export namespace Firelord {
 		? FirelordFirestore.Timestamp | Date
 		: T
 
-	// solve "Type instantiation is excessively deep and possibly infinite" error
-	// type ReadDeepConvert<T extends Record<string, unknown>> = {
-	// 	[K in keyof T]: ReadConverter<T[K]> extends Record<string, unknown>
-	// 		? ReadDeepConvert<ReadConverter<T[K]>>
-	// 		: ReadConverter<T[K]>
-	// }
-	// ! for some reason this does not work, WHY
-	// type WriteDeepConvert<T extends Record<string, unknown>> = {
-	// 	[K in keyof T]: WriteConverter<T[K]> extends Record<string, unknown>
-	// 		? WriteDeepConvert<WriteConverter<T[K]>>
-	// 		: WriteConverter<T[K]>
-	// }
-
-	// ! for some reason this does not work. WHY
-	// type CompareDeepConvert<T extends Record<string, unknown>> = {
-	// 	[K in keyof T]: CompareConverter<T[K]> extends Record<string, unknown>
-	// 		? CompareDeepConvert<CompareConverter<T[K]>>
-	// 		: CompareConverter<T[K]>
-	// }
-
 	export type ReadWriteCreator<
 		B extends { [index: string]: unknown },
 		C extends string,
@@ -153,15 +148,19 @@ export namespace Firelord {
 		}
 	> = {
 		base: B
-		read: ReadConverter<B> & {
-			[index in keyof FirelordFirestore.CreatedUpdatedRead]: FirelordFirestore.CreatedUpdatedRead[index]
-		}
+		read: DeepRequired<
+			ReadConverter<B> & {
+				[index in keyof FirelordFirestore.CreatedUpdatedRead]: FirelordFirestore.CreatedUpdatedRead[index]
+			}
+		>
 		// so it looks more explicit in typescript hint
-		writeNested: {
-			[J in keyof B]: WriteConverter<B[J]>
-		} & {
-			[index in keyof FirelordFirestore.CreatedUpdatedWrite]: FirelordFirestore.CreatedUpdatedWrite[index]
-		}
+		writeNested: DeepRequired<
+			{
+				[J in keyof B]: WriteConverter<B[J]>
+			} & {
+				[index in keyof FirelordFirestore.CreatedUpdatedWrite]: FirelordFirestore.CreatedUpdatedWrite[index]
+			}
+		>
 		write: {
 			[J in keyof FlattenObject<B>]: WriteConverter<FlattenObject<B>[J]>
 		} & {