+++ src/index_.ts
@@ -47,10 +47,12 @@ export type firelord = (firestore: FirelordFirestore.Firestore) => <
 				error?: ((error: Error) => void) | undefined
 			) => () => void
 			create: (
-				data: OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
+				data: OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
 			) => Promise<FirelordFirestore.WriteResult>
 			set: <
-				J_1 extends Partial<OmitKeys<T['write'], 'createdAt' | 'updatedAt'>>,
+				J_1 extends Partial<
+					OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
+				>,
 				Z extends {
 					merge?: true | undefined
 					mergeField?:
@@ -61,10 +63,10 @@ export type firelord = (firestore: FirelordFirestore.Firestore) => <
 				data: J_1 extends never
 					? J_1
 					: Z extends undefined
-					? OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
+					? OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
 					: Z['merge'] extends true
 					? PartialNoImplicitUndefinedAndNoExtraMember<
-							OmitKeys<T['write'], 'createdAt' | 'updatedAt'>,
+							OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>,
 							J_1
 					  >
 					: Z['mergeField'] extends Exclude<
@@ -72,10 +74,10 @@ export type firelord = (firestore: FirelordFirestore.Firestore) => <
 							'createdAt' | 'updatedAt'
 					  >[]
 					? PartialNoImplicitUndefinedAndNoExtraMember<
-							OmitKeys<T['write'], 'createdAt' | 'updatedAt'>,
+							OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>,
 							J_1
 					  >
-					: OmitKeys<T['write'], 'createdAt' | 'updatedAt'>,
+					: OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>,
 				options?: Z | undefined
 			) => Promise<FirelordFirestore.WriteResult>
 			update: <
@@ -104,17 +106,17 @@ export type firelord = (firestore: FirelordFirestore.Firestore) => <
 						  >
 				) => FirelordFirestore.WriteBatch
 				create: (
-					data: OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
+					data: OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
 				) => FirelordFirestore.WriteBatch
 			}
 			runTransaction: (
 				callback: (transaction: {
 					create: (
-						data: OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
+						data: OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
 					) => FirelordFirestore.Transaction
 					set: <
 						J_4 extends Partial<
-							OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
+							OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
 						>,
 						Z_1 extends {
 							merge?: true | undefined
@@ -126,11 +128,21 @@ export type firelord = (firestore: FirelordFirestore.Firestore) => <
 						data: J_4 extends never
 							? J_4
 							: Z_1 extends undefined
-							? OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
-							: PartialNoImplicitUndefinedAndNoExtraMember<
-									OmitKeys<T['write'], 'createdAt' | 'updatedAt'>,
+							? OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
+							: Z_1['merge'] extends true
+							? PartialNoImplicitUndefinedAndNoExtraMember<
+									OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>,
 									J_4
-							  >,
+							  >
+							: Z_1['mergeField'] extends Exclude<
+									keyof T['write'],
+									'createdAt' | 'updatedAt'
+							  >[]
+							? PartialNoImplicitUndefinedAndNoExtraMember<
+									OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>,
+									J_4
+							  >
+							: OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>,
 						options?: Z_1 | undefined
 					) => FirelordFirestore.Transaction
 					update: <
@@ -155,10 +167,10 @@ export type firelord = (firestore: FirelordFirestore.Firestore) => <
 			) => void
 		}
 		add: (
-			data: OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
+			data: OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
 		) => Promise<
 			FirelordFirestore.DocumentReference<
-				OmitKeys<T['write'], 'createdAt' | 'updatedAt'>
+				OmitKeys<T['writeNested'], 'createdAt' | 'updatedAt'>
 			>
 		>
 	} & ReturnType<